Почти закончил PizzaDetailScreen, пока убрал логику с CatalogScreen

+++ app/src/main/java/com/example/pizza_shift_2025/data/model/IngredientModel.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IngredientModel(
     val name: IngredientTypeModel,
-    val cost: Double,
+    val cost: Int,
     val img: String
 
 )
